cleanup(angular): refactor library generator to use nx devkit
Refactor Library Generator to use Nx Devkit
