+++ packages/angular/generators.ts
@@ -2,5 +2,6 @@ export * from './src/generators/convert-tslint-to-eslint/convert-tslint-to-eslin
 export * from './src/generators/karma/karma';
 export * from './src/generators/karma-project/karma-project';
 export * from './src/generators/move/move';
+export * from './src/generators/library/library';
 export * from './src/generators/storybook-configuration/storybook-configuration';
 export * from './src/schematics/generators';
