+++ packages/angular/src/generators/library/files/lib/README.md
