+++ packages/angular/src/generators/library/files/lib/tsconfig.json__tpl__
