+++ packages/angular/src/generators/library/lib/normalize-options.ts
@@ -1,6 +1,4 @@
-import { Tree } from '@angular-devkit/schematics';
-import { getNpmScope } from '@nrwl/workspace';
-import { libsDir } from '@nrwl/workspace/src/utils/ast-utils';
+import { getWorkspaceLayout, Tree } from '@nrwl/devkit';
 import { Schema } from '../schema';
 import { NormalizedSchema } from './normalized-schema';
 import { names } from '@nrwl/devkit';
@@ -14,16 +12,18 @@ export function normalizeOptions(
     ? `${names(options.directory).fileName}/${name}`
     : name;
 
+  const { libsDir, npmScope } = getWorkspaceLayout(host);
+
   const projectName = projectDirectory.replace(new RegExp('/', 'g'), '-');
   const fileName = options.simpleModuleName ? name : projectName;
-  const projectRoot = `${libsDir(host)}/${projectDirectory}`;
+  const projectRoot = `${libsDir}/${projectDirectory}`;
 
   const moduleName = `${names(fileName).className}Module`;
   const parsedTags = options.tags
     ? options.tags.split(',').map((s) => s.trim())
     : [];
   const modulePath = `${projectRoot}/src/lib/${fileName}.module.ts`;
-  const defaultPrefix = getNpmScope(host);
+  const defaultPrefix = npmScope;
 
   const importPath =
     options.importPath || `@${defaultPrefix}/${projectDirectory}`;
