+++ packages/angular/src/generators/library/lib/normalized-schema.ts
