+++ packages/angular/src/generators/library/lib/update-lib-package-npm-scope.ts
@@ -1,0 +1,12 @@
+import { Tree, updateJson } from '@nrwl/devkit';
+import { NormalizedSchema } from './normalized-schema';
+
+export function updateLibPackageNpmScope(
+  host: Tree,
+  options: NormalizedSchema
+) {
+  return updateJson(host, `${options.projectRoot}/package.json`, (json) => {
+    json.name = options.importPath;
+    return json;
+  });
+}
