+++ packages/angular/src/generators/library/library.compat.ts
@@ -1,0 +1,4 @@
+import { convertNxGenerator } from '@nrwl/devkit';
+import library from './library';
+
+export const librarySchematic = convertNxGenerator(library);
