+++ packages/angular/src/generators/library/schema.d.ts
