+++ packages/angular/src/generators/library/schema.json
@@ -1,8 +1,9 @@
 {
   "$schema": "http://json-schema.org/schema",
-  "$id": "SchematicsAngularLibrary",
+  "$id": "GeneratorAngularLibrary",
   "title": "Create a library",
   "type": "object",
+  "cli": "nx",
   "properties": {
     "name": {
       "type": "string",
