+++ packages/angular/src/schematics/generators.ts
@@ -1,3 +1,2 @@
 export { applicationGenerator } from './application/application';
-export { libraryGenerator } from './library/library';
 export { ngrxGenerator } from './ngrx/ngrx';
