+++ packages/workspace/src/generators/move/lib/move-project.ts
@@ -15,11 +15,8 @@ export function moveProject(
 ) {
   const destination = getDestination(tree, schema, project);
   visitNotIgnoredFiles(tree, project.root, (file) => {
-    // This is a rename but Angular Devkit isn't capable of writing to a file after it's renamed so this is a workaround
-    const content = tree.read(file);
     const relativeFromOriginalSource = relative(project.root, file);
     const newFilePath = join(destination, relativeFromOriginalSource);
-    tree.write(newFilePath, content);
-    tree.delete(file);
+    tree.rename(file, newFilePath);
   });
 }
